Add enable/disable support for hosts entries
Refactored HostModel to use QAbstractTableModel and added an 'enabled' property to each host entry. Hosts can now be enabled or disabled via a checkbox in the first column of the table, and this state is saved to and loaded from the hosts file using a '#!' prefix for disabled entries. Updated the UI to reflect these changes and improved column resizing for the checkbox column. Also updated the application window title.